Adds panels wrapper
